All the Ids properties of all the classes are now of type int  and now we don't have ProductId , ArtistId of GenreId in our classes Cd, Instrument and Score . They have now a property Product Product. Now we seed with anoniem class
Property ProductType and ProductTypeId deleted from the classes Cd, Instrument and Score becouse we have them already in the class Product

+++ Maracas.Lib/Models/Client.cs
@@ -6,15 +6,11 @@ namespace Maracas.Lib.Models
 {
     public class Client
     {
-        public long Id { get; set; }
-
+        public int Id { get; set; }
         public string Name { get; set; }
-
         public string Address { get; set; }
-
         public string City { get; set; }
         public string Email { get; set; }
-
         public List<Order> Orders { get; set; }
