+++ Maracas.Lib/Models/Genre.cs
@@ -6,14 +6,10 @@ namespace Maracas.Lib.Models
 {
     public class Genre
     {
-        public long Id { get; set; }
-
+        public int Id { get; set; }
         public string Name { get; set; }
-
         public List<Artist> Artists { get; set; }
-
         public List<Cd> Cds  { get; set; }
-
         public List <Score> Scores { get; set; }
 
     }
