+++ Maracas.Lib/Models/ProductType.cs
@@ -6,9 +6,8 @@ namespace Maracas.Lib.Models
 {
     public class ProductType
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Type { get; set; }
-
 
     }
 }
